refactor(components): [tabs] use JSX in Unit test
